feat(perusahaan mmea): pembuatan fitur create pembuatan fitur create data perusahaan cukai mmea perbaikan form filter period pada halaman master sbp dan perusahaan cukai ht + hptl

+++ database/migrations/2023_12_06_040723_create_perusahaan_mmea_table.php
@@ -30,8 +30,8 @@ return new class extends Migration
             $table->string('nama_perusahaan', 50);
             $table->string('nppbkc', 100);
             $table->integer('jumlah_dokumen')->unsigned();
-            $table->integer('jumlah_liter')->unsigned();
-            $table->decimal('jumlah_cukai', 15, 2)->unsigned();
+            $table->decimal('jumlah_liter', 17, 2)->unsigned();
+            $table->decimal('jumlah_cukai', 17, 2)->unsigned();
             $table->date('tanggal_input');
             $table->timestamps();
             $table->softDeletes();
